requirements.txt
numpy
pandas
pyarrow
fastparquet
requests
boto3
python-dotenv
